Promote WebGL v1 since it is widely supported by major browsers
