+++ src/main/scala/org/scalajs/dom/webgl/package.scala
@@ -1,9 +1,22 @@
-package org.scalajs.dom.experimental
+package org.scalajs.dom
 
 import org.scalajs.dom.raw.WebGLRenderingContext
 import scala.scalajs.js
 
 package object webgl {
+  type ActiveInfo = raw.WebGLActiveInfo
+  type Buffer = raw.WebGLBuffer
+  type ContextAttributes = raw.WebGLContextAttributes
+  type Framebuffer = raw.WebGLFramebuffer
+  type Program = raw.WebGLProgram
+  type Renderbuffer = raw.WebGLRenderbuffer
+  type RenderingContext = raw.WebGLRenderingContext
+  type Shader = raw.WebGLShader
+  type ShaderPrecisionFormat = raw.WebGLShaderPrecisionFormat
+  type Texture = raw.WebGLTexture
+  type UniformLocation = raw.WebGLUniformLocation
+  @inline def RenderingContext = raw.WebGLRenderingContext
+
   implicit class WebGLRenderingContextOps(val webGL: WebGLRenderingContext)
       extends AnyVal {
